pyproject.toml
[project]
name = "hackaton-ai-2026"
version = "0.1.0"
description = "AI service for the Voice from the Archive hackathon"
readme = "docs/get_info_implementation_plan.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "jinja2>=3.1.0",
  "openai>=1.99.0",
  "pydantic-settings>=2.4.0",
  "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

